requirements.txt
httpx>=0.27.0
fastapi>=0.110.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0
loguru>=0.7.0
alpaca-py>=0.20.0
pytz>=2024.1
